moved RevisionedPersistenceEntity to devon4j-basic (was accidentally forgotten on refactoring)
