+++ modules/basic/src/main/java/com/devonfw/module/basic/common/api/entity/RevisionedPersistenceEntity.java
@@ -1,9 +1,6 @@
 /* Copyright (c) The m-m-m Team, Licensed under the Apache License, Version 2.0
  * http://www.apache.org/licenses/LICENSE-2.0 */
-package com.devonfw.module.jpa.dataaccess.api;
-
-import com.devonfw.module.basic.common.api.entity.PersistenceEntity;
-import com.devonfw.module.basic.common.api.entity.RevisionedEntity;
+package com.devonfw.module.basic.common.api.entity;
 
 /**
  * This is the interface for a {@link RevisionedEntity revisioned} {@link PersistenceEntity persistence entity}.
