+++ modules/jpa-envers/src/main/java/com/devonfw/module/jpa/dataaccess/api/RevisionedDao.java
@@ -1,6 +1,6 @@
 package com.devonfw.module.jpa.dataaccess.api;
 
-import com.devonfw.module.jpa.dataaccess.api.RevisionedPersistenceEntity;
+import com.devonfw.module.basic.common.api.entity.RevisionedPersistenceEntity;
 
 /**
  * This is a simplified variant of {@link GenericRevisionedDao} for the suggested and common case that you have a
